backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
sqlmodel
python-jose[cryptography]==3.3.0
python-dotenv==1.0.1
pydantic>=2.0,<3.0
psycopg2-binary
google-generativeai
